Map typings to STAN's
